requirements.txt
pandas>=1.4.0
openpyxl>=3.0.10
SQLAlchemy>=1.4.0
pymysql>=1.0.2
pydantic>=1.10.0
python-dotenv>=0.21.0
tqdm>=4.64.0
mysql-connector-python>=8.0.33
